[PDB] read/dump fixups from PDB dbi stream
Summary:
Binaries linked with `/DEBUGTYPE:CV,FIXUP` have a fixup substream present in the generated PDB's dbi.
This is essentially the equivalent of GCC's `--emit-relocs`. The fixup stream contains a list of all static
relocations performed by the linker. Each fixup entry uses the `XFIXUP_DATA` structure found in the cvdump
reference implementation.

Changes to llvm-pdbutil's dump tool can dump this information with `-fixups` or `-all`. The `DbiStream` public
API provides access to this stream array for tools that may be interested in cross-referencing symbols with
their usages.

Subscribers: hiraditya, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D83014